+++ modules/common.py
@@ -2,7 +2,6 @@ import tempfile
 from datetime import datetime
 
 import faicons
-from bs4 import BeautifulSoup
 from matplotlib import colormaps
 from shiny import Inputs, Outputs, Session, module, reactive, render, ui
 from synthetmic.data.utils import SynthetMicData
@@ -159,21 +158,21 @@ def server(
         return ui.card(
             ui.card_header("Common plot options"),
             ui.tags.div(
-                comps.create_selection(
+                comps.selection(
                     id="view",
                     label="Choose a diagram to view",
                     choices=view_choices,
                     selected=DiagramView.FULL,
                     width="100%",
                 ),
-                comps.create_selection(
+                comps.selection(
                     id="colorby",
                     label="Color by",
                     choices=[c for c in Colorby],
                     selected=PLOT_DEFAULTS["colorby"],
                     width="100%",
                 ),
-                comps.create_selection(
+                comps.selection(
                     id="colormap",
                     label="Choose a colormap",
                     choices=sorted(list(colormaps)),
@@ -189,7 +188,7 @@ def server(
                     ticks=True,
                     width="100%",
                 ),
-                comps.create_input_action_button(
+                comps.input_action_button(
                     id="reset_plot_options",
                     label="Reset common plot options to defaults",
                     icon="gear",
@@ -261,7 +260,7 @@ def server(
 
             case DiagramView.SLICE:
                 opts += (
-                    comps.create_selection(
+                    comps.selection(
                         id="slice_normal",
                         label="Choose a normal for slicing",
                         choices=list(utils.COORDINATES[: len(data.domain)]),
@@ -272,7 +271,7 @@ def server(
                 )
             case DiagramView.CLIP:
                 opts += (
-                    comps.create_selection(
+                    comps.selection(
                         id="clip_normal",
                         label="Choose a normal for clipping",
                         choices=list(utils.COORDINATES[: len(data.domain)]),
@@ -305,22 +304,19 @@ def server(
     @render.ui
     def metrics_tab_card() -> ui.Tag | None:
         if input.view() == DiagramView.CLIP:
-            return ui.card(
-                ui.card_header(
-                    faicons.icon_svg("circle-exclamation", fill=FILL_COLOUR),
-                    "A note on metrics",
-                ),
+            return ui.help_text(
                 ui.markdown(
                     """
-                    Metrics are not available for clip view. You can only view the clipped microstructure
-                    by selecting the **Microstructure** tab.
+                    Metrics are not available for clip view. You can
+                    only view the clipped microstructure
+                    by selecting the **Microstructure** button.
                     """
                 ),
             )
 
         metrics = _calculate_metrics()
         if isinstance(metrics, Exception):
-            comps.create_error_notification(str(metrics))
+            comps.error_notification(str(metrics))
             return
 
         @render.plot
@@ -354,7 +350,7 @@ def server(
                     """
                 )
             ),
-            comps.create_download_button(
+            comps.download_button(
                 id="download_metrics",
                 label="Yes, download the current metrics",
             ),
@@ -419,7 +415,7 @@ def server(
     def diagram_tab_card() -> ui.Tag | None:
         diagram = _calculate_diagram()
         if isinstance(diagram, Exception):
-            comps.create_error_notification(str(diagram))
+            comps.error_notification(str(diagram))
             return
 
         @render.ui
@@ -429,9 +425,10 @@ def server(
                 f.seek(0)
                 html = f.read().decode("utf-8")
 
-            soup = BeautifulSoup(html, "html.parser")
-            html = str(soup.body)
-            return ui.HTML(html)
+            # return ui.HTML(html)
+            return ui.tags.iframe(
+                srcdoc=html, style="width:100%; height:100%; border:none;"
+            )
 
         @render.download(
             filename=lambda: f"synthetmic-gui-output-{datetime.now().strftime('%Y-%m-%d-%H-%M-%S')}.zip",
@@ -456,7 +453,7 @@ def server(
                     "All outputs will be downloaded in .zip format which can be easily unzipped for futher processing."
                 )
             ),
-            comps.create_download_button(
+            comps.download_button(
                 id="download_diagram",
                 label="Yes, download the current diagram and properties",
             ),
