+++ pyproject.toml
@@ -16,7 +16,6 @@ dependencies = [
     "shinywidgets>=0.6.2",
     "pyvista==0.46.5",
     "trame-vtk==2.9.1",
-    "beautifulsoup4==4.14.3",
 ]
 
 [project.optional-dependencies]
