+++ www/ga/analytics.html
@@ -21,7 +21,7 @@
 
 <script>
     // For download buttons
-    $(document).on('click', '#common-download_diagram', function () {
+    $(document).on('click', '#main-common-download_diagram', function () {
         gtag(
             'event',
             'diagram_download',
@@ -33,7 +33,7 @@
         );
     });
 
-    $(document).on('click', '#common-download_metrics', function () {
+    $(document).on('click', '#main-common-download_metrics', function () {
         gtag(
             'event',
             'metrics_download',
