+++ www/layout.css
@@ -2,7 +2,7 @@
     justify-content: center !important;
 }
 
-.modal-dialog {
-    margin-top: 20px !important;
-    overflow-y: hidden !important;
+.popover {
+    max-width: 350px !important;
+    width: 350px !important;
 }
