+++ www/style.css
@@ -1,0 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Atkinson+Hyperlegible:ital,wght@0,400;0,700;1,400;1,700&display=swap');
+
+* {
+    font-family: 'Atkinson Hyperlegible', sans-serif;
+}
